% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{data Earthquake NOAA Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{raw}{contains the original raw dataset}
}
\value{
This function returns the clean version of the dataset
}
\description{
function: eq_clean_data()
}
\details{
\code{eq_clean_data()} that takes raw NOAA data frame and returns a clean data frame.
The clean data frame should have the following:
A date column created by uniting the year, month, day and converting it to the \code{Date} class.
\code{LATITUDE} and \code{LONGITUDE} columns converted to numeric class.
}
\examples{

\dontrun{
data <- readr::read_delim("NOAAearthquakes.txt", delim = "\\t")
clean_data <- eq_clean_data(data)
}


}
