% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean
Parse LOCATION_NAME field of NOAA dataset}
\usage{
eq_location_clean(locationtoclean)
}
\arguments{
\item{data}{A data frame with raw data obtained from NOAA website}

\item{location_name}{represents the detailed information about location}
}
\value{
A data frame with cleaned LOCATION_NAME column

This function returns the clean version of the LOCATION_NAME
}
\description{
eq_location_clean
Parse LOCATION_NAME field of NOAA dataset
}
\examples{
\dontrun{
library(readr)
data <- readr::read_delim(("extdata","earthquake.txt", delim = "\\t")
data <- eq_location_clean(data)
}

}
