% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A Dataframe Object containing the data to plot (Must include a LATITUDE and LONGITUDE colums).}

\item{annot_col}{Column of `data` depicting the markers that will be used in the map (upon clicking).}
}
\value{
This function adds the map as well as the datapoints, positioned at their respective Lat and Long
coordinates, into the current graphics device. Furthermore, it add popup markers for each datapoint depending
on the \code{annot_col} arguments.
}
\description{
This function takes a \code{data} argument that must include LATITUDE and LONGITUDE columns, and displays
its data points in a leaflet map. The points are depicted as circles, with popup annotations being displayed
upon clicking. These annotations depend on the \code{annot_col} argument.
}
\examples{

\dontrun{

readr::read_delim("data/NOAAearthquakes.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")

}

}
