% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Mapping argument to the ggplot layer function.}

\item{data}{A Dataframe Object containing the data to plot.}

\item{stat}{Stat argument to the ggplot layer function.}

\item{position}{Position argument to the ggplot layer function.}

\item{na.rm}{na.rm argument to the ggplot layer function.}

\item{show.legend}{show.legend argument to the ggplot layer function.}

\item{inherit.aes}{inherit.aes argument to the ggplot layer function.}

\item{...}{Extra Params.

*Note: All of the optional arguments above can be applied as columns of the dataset (inside the function aes),
or as individual values (outside of the function aes).}
}
\value{
This function adds the lineGrobs and pointGrob into the current graphics device.
}
\description{
This layer function displays the dates of Earthquakes (coming from your \code{data}) by
presenting each Earthquake, on a straight line, as a point of varying sizes depending on
their Ritcher scale value (based on the \code{size} variable). The Earthquakes are color
coded by the death toll it caused (\code{colour}). If the \code{y} variable is supplied,
the Earthquakes will be displayed in multiple horizontal lines, each line represnting
the quakes for each Country.
}
\section{Parameters to Geom}{


\code{data}: A Dataframe Object containing the data to plot. \cr \cr
\code{x}: Column of `data` depicting the x-coordinate of where each of the points will be located. \cr \cr
\code{size}: OPTIONAL. Column of `data` depicting size of each point in the pointsGrob displayed (if
missing, the first \code{nmax} labels will be displayed). \cr \cr
\code{y}: OPTIONAL. Column of `data` depicting the y-coordinate of where each of the points will be located. \cr \cr
\code{colour}: OPTIONAL. Column of `data` depicting the colour of each point in the pointGrob displayed. \cr \cr
\code{shape}: OPTIONAL. Column of 'data' depicting the shape of each of the points. \cr \cr
\code{alpha}: OPTIONAL. Column of `data` which will be used to apply the alpha to each point. \cr \cr
\code{fill}: OPTIONAL. Color 'string' (added outside of aes) to modify the default fill of the points (before applying \code{colour}). \cr \cr
\code{stroke}: OPTIONAL. Stroke of the image (need not be used). \cr \cr

*Note: All of the optional arguments above can be applied as columns of the dataset (inside the function aes),
or as individual values (outside of the function aes).
}

\examples{

\dontrun{

## Data Setup

dataset = readr::read_delim("data/NOAAearthquakes.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)

## Function Call

ggplot2::ggplot(dataset, ggplot2::aes(DATE,
                                      colour = TOTAL_DEATHS,
                                      y = COUNTRY,
                                      size = as.numeric(EQ_PRIMARY) )) +
  geom_timeline() +
  ggplot2::guides(size = ggplot2::guide_legend(title = "Ritcher scale value")) +
  ggplot2::scale_colour_continuous(name = "# of DEATHS") +
  ggplot2::theme_classic()

}

}
