% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Mapping argument to the ggplot layer function.}

\item{data}{A Dataframe Object containing the data to plot.}

\item{stat}{Stat argument to the ggplot layer function.}

\item{position}{Position argument to the ggplot layer function.}

\item{na.rm}{na.rm argument to the ggplot layer function.}

\item{show.legend}{show.legend argument to the ggplot layer function.}

\item{inherit.aes}{inherit.aes argument to the ggplot layer function.}

\item{...}{Extra Params.}
}
\value{
This function adds the lineGrobs and textGrobs into the current graphics device.
}
\description{
This layer function displays the labels of Earthquakes (coming from your \code{label}) by
presenting each label, rotated 45 degrees, as text.
The top \code{nmax} labels, based on the \code{size} column, will be displayed on top of a vertical line.
This vertical line has one end located at the Earthquakes' x-coordinates (dates), and the other end under the label.
If given a \code{y} argument, the vertical lines will be displayed in two groups at each of the y-coordinates; the
top \code{nmax} labels will be displayed for EACH y-coordinate.
}
\section{Parameters to Geom}{


\code{data}: A Dataframe Object containing the data to plot. \cr \cr
\code{x}: Column of `data` depicting the x-coordinate of where each of the labels will be located. \cr \cr
\code{label}: Column of `data` depicting the text labels that will be displayed. \cr \cr
\code{size}: Column of `data` depicting size of the Earthquake, which will be used to decide the top \code{nmax} labels displayed. \cr \cr
\code{y}: OPTIONAL. Column of `data` depicting the y-coordinate of each label displayed. \cr \cr
\code{nmax}: OPTIONAL. Column of 'data' depicting how many top \code{size}d labels will be displayed. \cr \cr
\code{alpha}: OPTIONAL. Alpha numeric value (0 to 1, added outside of aes) which will be used to apply the alpha to the vertical line. \cr \cr
\code{colour}: OPTIONAL. Color 'string' (added outside of aes) to modify the default color of the lines. \cr \cr
\code{stroke}: OPTIONAL. Stroke of the image (need not be used). \cr \cr

*Note: All of the optional arguments above can be applied as columns of the dataset (inside the function aes),
or as individual values (outside of the function aes).
}

\examples{

\dontrun{

## Data Setup

dataset = readr::read_delim("data/NOAAearthquakes.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)

## Function Call

ggplot2::ggplot(dataset, ggplot2::aes(DATE,
                                      colour = TOTAL_DEATHS,
                                      y = COUNTRY,
                                      size = as.numeric(EQ_PRIMARY) )) +
  geom_timeline() +
  geom_timeline_label(ggplot2::aes(label = dataset$LOCATION_NAME)) +
  ggplot2::guides(size = ggplot2::guide_legend(title = "Ritcher scale value")) +
  ggplot2::scale_colour_continuous(name = "# of DEATHS") +
  ggplot2::theme_classic()

}

}
