% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\value{
This function return a Geom* object containing the line and point Grobs responsible for rendering the images into the plot.
}
\description{
This Geom returns a ggproto object made of a gList of lineGrobs and pointGrobs. Each lineGrob is a straight horizontal line ranging
from the minimum x-value to maximum x-value. This horizontal line is done for each unique y-value.
The Point grobs are colored based on the \code{colour} argument. There will be one point per row.
}
\section{Parameters to Geom}{


\code{data}: A Dataframe Object containing the data to plot. \cr \cr
\code{x}: Column of `data` depicting the x-coordinate of where each of the points will be located. \cr \cr
\code{size}: OPTIONAL. Column of `data` depicting size of each point in the pointsGrob displayed (if
missing, the first \code{nmax} labels will be displayed). \cr \cr
\code{y}: OPTIONAL. Column of `data` depicting the y-coordinate of where each of the points will be located. \cr \cr
\code{colour}: OPTIONAL. Column of `data` depicting the colour of each point in the pointGrob displayed. \cr \cr
\code{shape}: OPTIONAL. Column of 'data' depicting the shape of each of the points. \cr \cr
\code{alpha}: OPTIONAL. Column of `data` which will be used to apply the alpha to each point. \cr \cr
\code{fill}: OPTIONAL. Color 'string' (added outside of aes) to modify the default fill of the points (before applying \code{colour}). \cr \cr
\code{stroke}: OPTIONAL. Stroke of the image (need not be used). \cr \cr

*Note: All of the optional arguments above can be applied as columns of the dataset (inside the function aes),
or as individual values (outside of the function aes).
}

\examples{

\dontrun{

## Data Setup

dataset = readr::read_delim("NOAAearthquakes.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)

## Function Call

ggplot2::ggplot(dataset, ggplot2::aes(DATE,
                                      colour = TOTAL_DEATHS,
                                      y = COUNTRY,
                                      size = as.numeric(EQ_PRIMARY) )) +
  geom_timeline() +
  ggplot2::guides(size = ggplot2::guide_legend(title = "Ritcher scale value")) +
  ggplot2::scale_colour_continuous(name = "# of DEATHS") +
  ggplot2::theme_classic()

}

}
\keyword{datasets}
