% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomTimelineLabel
}
\value{
This function return a Geom* object containing the line and text Grobs responsible for rendering the images into the plot.
}
\description{
This Geom returns a ggproto object made of a gList of lineGrobs and textGrobs. Each lineGrob is a vertical line located
at the points' x-axis. The height of the lines depends on how many y-degrees are given.
The Text grobs writes the text found on the \code{label} argument. There will be one label per row, although the rows are filtered
for only the top \code{nmax} labels, based on their \code{size} column. The text is rotated 45 degrees, and it's positioned slightly
above the top end of the lineGrob.
The Geom allows you to label the higher-valued (based on \code{size}) points in your dataset.
}
\section{Parameters}{


\code{data}: A Dataframe Object containing the data to plot. \cr \cr
\code{x}: Column of `data` depicting the x-coordinate of where each of the labels will be located. \cr \cr
\code{label}: Column of `data` depicting the text labels that will be displayed. \cr \cr
\code{size}: Column of `data` depicting size of the Earthquake, which will be used to decide the top \code{nmax} labels displayed. \cr \cr
\code{y}: OPTIONAL. Column of `data` depicting the y-coordinate of each label displayed. \cr \cr
\code{nmax}: OPTIONAL. Column of 'data' depicting how many top \code{size}d labels will be displayed. \cr \cr
\code{alpha}: OPTIONAL. Alpha numeric value (0 to 1, added outside of aes) which will be used to apply the alpha to the vertical line. \cr \cr
\code{colour}: OPTIONAL. Color 'string' (added outside of aes) to modify the default color of the lines. \cr \cr
\code{stroke}: OPTIONAL. Stroke of the image (need not be used). \cr \cr

*Note: All of the optional arguments above can be applied as columns of the dataset (inside the function aes),
or as individual values (outside of the function aes).
}

\examples{

\dontrun{

## Data Setup

dataset = readr::read_delim("data/NOAAearthquakes.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)

## Function Call

ggplot2::ggplot(dataset, ggplot2::aes(DATE,
                                      colour = TOTAL_DEATHS,
                                      y = COUNTRY,
                                      size = as.numeric(EQ_PRIMARY) )) +
  geom_timeline() +
  geom_timeline_label(ggplot2::aes(label = dataset$LOCATION_NAME)) +
  ggplot2::guides(size = ggplot2::guide_legend(title = "Ritcher scale value")) +
  ggplot2::scale_colour_continuous(name = "# of DEATHS") +
  ggplot2::theme_classic()

}

}
\keyword{datasets}
