% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{function: eq_clean_data()}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{Dataset depicting the raw NOAA data frame.}
}
\value{
it returns the data frame, cleaned and with correct dates (with date class)
}
\description{
This function returns the raw NOAA data frame with a new \code{DATE} column (of class date) added
by combining the MONTH, YEAR, and DAY fields (MONTHS and DAYS with NA are replaced to 01).
Furthermore, the rows with NA YEAR, or YEAR < 0 are removed. The LATITUDE and LONGITUDE fields
are converted to numeric type, and the LOCATION_NAME fixing (from the eq_location_clean function)
is implemented.
}
\examples{

\dontrun{

dataset <- read_delim("data/NOAAearthquakes.txt", delim = "\\t")
dataset <- eq_clean_data(dataset)

}
}
