% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the raw NOAA earquake data}
\usage{
eq_clean_data(cleandata)
}
\arguments{
\item{datatoclean}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned date, latitude and longitude numerical columns
}
\description{
Clean the raw NOAA earquake data
}
\examples{
\dontrun{
data <- readr::read_delim("earthquake.txt", delim = "\\t")
data <- eq_clean_data(data)
}

}
