% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{More interesting pop-ups for the interactive map used with the eq_map() function}
\usage{
eq_create_label(mapdata)
}
\arguments{
\item{mapdata}{A cleaned data frame with data obtained from NOAA website}
}
\value{
An HTML label that can be used as the annotation text in the leaflet map.
}
\description{
More interesting pop-ups for the interactive map used with the eq_map() function
}
\examples{
\dontrun{
readr::read_delim(("extdata","earthquake.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}

}
