% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{Dataset depicting the raw NOAA data frame.}
}
\value{
This function returns the raw NOAA data frame with the LOCATION_NAME columns cleaned up.
}
\description{
This function modifies the LOCATION_NAME column of a dataframe by separating the LOCATION_NAME
into multiple strings (separated by the character ':'), and selecting the last one. This step
transforms the LOCATION_NAME field from states and cities (towns) to just cities.
It also correctly formats the LOCATION_NAME so that it only has its first letter capitalized.
}
\examples{

\dontrun{

dataset <- read_delim("data/NOAAearthquakes.txt", delim = "\\t")
dataset <- eq_location_clean(dataset)

}

}
