% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Mapping the earthquake epicenters and providing some annotations from NOAA earthquake data}
\usage{
eq_map(mapdata, annot_col = "DATE")
}
\arguments{
\item{annot_col}{a string representing colume}

\item{dat}{an imput dataframe}
}
\value{
A map of the earthquakes epicenters and providing some annotations
}
\description{
This interactive map can show earquake information on popup
 accoring to given column
}
\examples{
\dontrun{
readr::read_delim(("extdata","earthquake.txt", delim = "\\t") \%>\%
eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
}

}
