% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Visualize the times at which earthquakes occur within certain countries}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\value{
The geom \code{geom_timeline} used with the \code{ggplot} function,is used to display time series events ranging from one
time point to another time point
}
\description{
Visualize the times at which earthquakes occur within certain countries
}
\examples{
# The data must be cleaned using the function \\code{eq_clean_data}, included in the package.
# Aesthetics can be specified in the \\code{ggplot} function or in \\code{geom_timeline} geom function
\dontrun{
data <- readr::read_delim(("extdata","earthquake.txt", delim = "\\t")
data <- eq_clean_data(data)
data \%>\%
dplyr::filter(COUNTRY == c("MEXICO","USA") & lubridate::year(date) >= 2010) \%>\%
ggplot(aes(x=DATE,y=COUNTRY,color=TOTAL_DEATHS,size=EQ_PRIMARY)) +
geom_timeline(alpha=.5) +
theme(legend.position="bottom", legend.box="horizontal", plot.title=element_text(hjust=0.5)) +
ggtitle("Earthquakes Visualization Tool") +
labs(size = "Richter scale value", color = "# deaths")
}
}
